+++ node/clients/payments.ts
@@ -2,13 +2,22 @@ import { ExternalClient, InstanceOptions, IOContext } from '@vtex/api'
 
 export default class AffirmDataSource extends ExternalClient {
   constructor(context: IOContext, options?: InstanceOptions) {
-    super(
-      `http://${context.workspace}--${context.account}.${process.env.VTEX_PUBLIC_ENDPOINT}/affirm/payment-provider/`,
-      context,
-      options
-    )
+    super(``, context, options)
   }
 
-  public getPaymentRequest = (qs: string) =>
-    this.http.get(`payments/${qs}/request`, { metric: 'payment-request' })
+  public getPaymentRequest = (
+    qs: string,
+    account: string = this.context.account,
+    workspace: string = this.context.workspace
+  ) => {
+    const base = this.baseUrl(account, workspace)
+
+    return this.http.get(`${base}/payments/${qs}/request`, {
+      metric: 'payment-request',
+    })
+  }
+
+  public baseUrl = (account: string, workspace: string) => {
+    return `http://${workspace}--${account}.${process.env.VTEX_PUBLIC_ENDPOINT}/affirm/payment-provider`
+  }
 }